Fixed ping-update being delayed when joining new server
When joining a server, the ping update is delayed until the tick interval is reached. However, the tick counter doesn't reset when the player joins a server/world, which means the next ping could be delayed by up to the full interval (e20 to 120 seconds).
To fix this, the ping should update shortly after connecting. The ping update iniside the init function is not needed.